+++ src/utils/systemReport.ts
@@ -15,7 +15,10 @@ import {
     deviceInfo as getDeviceInfo,
     productPageUrl,
 } from '../Device/deviceInfo/deviceInfo';
-import { getDeviceLibContext } from '../Device/deviceLibWrapper';
+import {
+    getDeviceLibContext,
+    getModuleVersion,
+} from '../Device/deviceLibWrapper';
 import logger from '../logging';
 import { Device } from '../state';
 import { getAppDataDir } from './appDirs';
@@ -52,9 +55,6 @@ const generalInfoReport = async () => {
         getDeviceLibContext()
     );
 
-    const nrfjprog = versions.find(v => v.moduleName === 'nrfjprog_dll');
-    const jlink = versions.find(v => v.moduleName === 'jlink_dll');
-
     return [
         `- System:     ${manufacturer} ${model}`,
         `- BIOS:       ${vendor} ${version}`,
@@ -73,8 +73,10 @@ const generalInfoReport = async () => {
         `    - node: ${node}`,
         `    - python: ${python}`,
         `    - python3: ${python3}`,
-        nrfjprog ? `    - nrfjprog: ${describeVersion(nrfjprog)}` : '',
-        jlink ? `    - jlink: ${describeVersion(jlink)}` : '',
+        `    - nrfjprog DLL: ${describeVersion(
+            getModuleVersion('jprog', versions)
+        )}`,
+        `    - JLink: ${describeVersion(getModuleVersion('jlink', versions))}`,
         '',
     ];
 };